Adds PaymentMethodData JSON converter
Adds a custom JSON converter for `PaymentMethodData` to handle deserialization of concrete payment method types based on the "type" field.

This allows the system to correctly instantiate `CreditCardPaymentMethodData`, `GooglePayPaymentMethodData`, or `PaypalPaymentMethodData` when deserializing a `PaymentMethodData` object from JSON.

Also, this updates the PayPal payment method definition to use the correct class.